LayoutTests/imported/w3c/web-platform-tests/css/css-multicol/multicol-fill-balance-025-expected.txt


FAIL Clearance after float with forced break inside assert_equals: expected 100 but got 74
